[01113] Move shortcut utilities to shared @/lib/shortcut module
